+++ modules/Grades/MassCreateAssignments.php
@@ -516,7 +516,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 	$link = [];
 
-	$link['TITLE']['link'] = 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=' . $_REQUEST['modfunc'];
+	$link['TITLE']['link'] = 'Modules.php?modname=' . $_REQUEST['modname'];
 
 	$link['TITLE']['variables'] = [ 'assignment_type' => 'TITLE_FOR_LINK' ];